requirements.txt
# Core packages
numpy==2.2.5
pandas==2.2.2
scipy==1.15.2
pillow==10.2.0
tqdm==4.66.4
pyyaml==6.0.1
requests==2.32.3
orjson==3.10.15
psutil==5.9.0
typing-extensions==4.9.0
packaging==24.0
python-dateutil==2.9.0
pydantic==2.6.4
pydantic-core==2.16.3

# PyTorch ecosystem
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.48.0
datasets==2.19.2
accelerate==0.34.0
peft==0.10.0
bitsandbytes==0.43.1
deepspeed==0.15.4
safetensors==0.4.5
einops==0.7.0
trl==0.8.6

# Computer Vision
opencv-python-headless==4.11.0.86
scikit-image==0.25.2
pymupdf==1.25.5
shapely==2.1.0
pyclipper==1.3.0.post6
easyocr==1.7.2
decord==0.6.0
av==14.1.0
ffmpy==0.5.0
qwen-vl-utils==0.0.11

# NLP
spacy==3.8.5
sentencepiece==0.2.0
tokenizers==0.21.0
langcodes==3.5.0
python-bidi==0.6.6
levenshtein==0.27.1
rapidfuzz==3.13.0
marisa-trie==1.2.1

# Web/API
fastapi==0.115.8
uvicorn==0.34.0
aiohttp==3.9.5
httpx==0.28.1
gradio==5.14.0
gradio-client==1.7.0

# Medical Imaging (optional)
pyxnat==1.6.3
nibabel==5.3.2
nipype==1.10.0

# Utilities
rich==13.7.1
wandb==0.16.6
fsspec==2024.3.1
filelock==3.13.1
protobuf==4.25.3
ninja==1.11.1.1

# CUDA (only needed if using GPU)
nvidia-ml-py==12.560.30
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvtx-cu12==12.4.127